feat: code highlighting has been pulled out into a composable and zero-docs code blocks now have a more legible light theme
