+++ packages/zero-docs/assets/scss/components/code.scss
@@ -48,45 +48,41 @@
 
 // ====================================================================== Themes
 html {
-  // ///////////////////////////////////////////////////// [theme] Atom One Dark
-  // https://raw.githubusercontent.com/highlightjs/highlight.js/main/src/styles/atom-one-dark.css
+  // //////////////////////////////////////////////////// [theme] Atom One Light
+  // https://raw.githubusercontent.com/highlightjs/highlight.js/main/src/styles/atom-one-light.css
   .code-block {
     /*
 
-    Atom One Dark by Daniel Gamage
-    Original One Dark Syntax theme from https://github.com/atom/one-dark-syntax
+    Atom One Light by Daniel Gamage
+    Original One Light Syntax theme from https://github.com/atom/one-light-syntax
 
-    base:    #282c34
-    mono-1:  #abb2bf
-    mono-2:  #818896
-    mono-3:  #5c6370
-    hue-1:   #56b6c2
-    hue-2:   #61aeee
-    hue-3:   #c678dd
-    hue-4:   #98c379
-    hue-5:   #e06c75
-    hue-5-2: #be5046
-    hue-6:   #d19a66
-    hue-6-2: #e6c07b
+    base:    #fafafa
+    mono-1:  #383a42
+    mono-2:  #686b77
+    mono-3:  #a0a1a7
+    hue-1:   #0184bb
+    hue-2:   #4078f2
+    hue-3:   #a626a4
+    hue-4:   #50a14f
+    hue-5:   #e45649
+    hue-5-2: #c91243
+    hue-6:   #986801
+    hue-6-2: #c18401
 
     */
-    color: #abb2bf;
-    transition: color 500ms;
 
-    span {
-      transition: color 500ms;
-    }
+    color: var(--code__block__color);
 
     .hljs-comment,
     .hljs-quote {
-      color: #5c6370;
+      color: #a0a1a7;
       font-style: italic;
     }
 
     .hljs-doctag,
     .hljs-keyword,
     .hljs-formula {
-      color: #c678dd;
+      color: #a626a4;
     }
 
     .hljs-section,
@@ -94,11 +90,11 @@ html {
     .hljs-selector-tag,
     .hljs-deletion,
     .hljs-subst {
-      color: #e06c75;
+      color: #e45649;
     }
 
     .hljs-literal {
-      color: #56b6c2;
+      color: #0184bb;
     }
 
     .hljs-string,
@@ -106,7 +102,7 @@ html {
     .hljs-addition,
     .hljs-attribute,
     .hljs-meta .hljs-string {
-      color: #98c379;
+      color: #50a14f;
     }
 
     .hljs-attr,
@@ -117,7 +113,7 @@ html {
     .hljs-selector-attr,
     .hljs-selector-pseudo,
     .hljs-number {
-      color: #d19a66;
+      color: #986801;
     }
 
     .hljs-symbol,
@@ -126,13 +122,13 @@ html {
     .hljs-meta,
     .hljs-selector-id,
     .hljs-title {
-      color: #61aeee;
+      color: #4078f2;
     }
 
     .hljs-built_in,
     .hljs-title.class_,
     .hljs-class .hljs-title {
-      color: #e6c07b;
+      color: #c18401;
     }
 
     .hljs-emphasis {
@@ -147,42 +143,46 @@ html {
       text-decoration: underline;
     }
   }
-  // //////////////////////////////////////////////////// [theme] Atom One Light
-  // https://raw.githubusercontent.com/highlightjs/highlight.js/main/src/styles/atom-one-light.css
+  // ///////////////////////////////////////////////////// [theme] Atom One Dark
+  // https://raw.githubusercontent.com/highlightjs/highlight.js/main/src/styles/atom-one-dark.css
   &.dark {
     .code-block {
+
       /*
 
-      Atom One Light by Daniel Gamage
-      Original One Light Syntax theme from https://github.com/atom/one-light-syntax
+      Atom One Dark by Daniel Gamage
+      Original One Dark Syntax theme from https://github.com/atom/one-dark-syntax
 
-      base:    #fafafa
-      mono-1:  #383a42
-      mono-2:  #686b77
-      mono-3:  #a0a1a7
-      hue-1:   #0184bb
-      hue-2:   #4078f2
-      hue-3:   #a626a4
-      hue-4:   #50a14f
-      hue-5:   #e45649
-      hue-5-2: #c91243
-      hue-6:   #986801
-      hue-6-2: #c18401
+      base:    #282c34
+      mono-1:  #abb2bf
+      mono-2:  #818896
+      mono-3:  #5c6370
+      hue-1:   #56b6c2
+      hue-2:   #61aeee
+      hue-3:   #c678dd
+      hue-4:   #98c379
+      hue-5:   #e06c75
+      hue-5-2: #be5046
+      hue-6:   #d19a66
+      hue-6-2: #e6c07b
 
       */
+      transition: color 500ms;
 
-      color: #D3D3D3;
+      span {
+        transition: color 500ms;
+      }
 
       .hljs-comment,
       .hljs-quote {
-        color: #a0a1a7;
+        color: #5c6370;
         font-style: italic;
       }
 
       .hljs-doctag,
       .hljs-keyword,
       .hljs-formula {
-        color: #a626a4;
+        color: #c678dd;
       }
 
       .hljs-section,
@@ -190,11 +190,11 @@ html {
       .hljs-selector-tag,
       .hljs-deletion,
       .hljs-subst {
-        color: #e45649;
+        color: #e06c75;
       }
 
       .hljs-literal {
-        color: #0184bb;
+        color: #56b6c2;
       }
 
       .hljs-string,
@@ -202,7 +202,7 @@ html {
       .hljs-addition,
       .hljs-attribute,
       .hljs-meta .hljs-string {
-        color: #50a14f;
+        color: #98c379;
       }
 
       .hljs-attr,
@@ -213,7 +213,7 @@ html {
       .hljs-selector-attr,
       .hljs-selector-pseudo,
       .hljs-number {
-        color: #986801;
+        color: #d19a66;
       }
 
       .hljs-symbol,
@@ -222,13 +222,13 @@ html {
       .hljs-meta,
       .hljs-selector-id,
       .hljs-title {
-        color: #4078f2;
+        color: #61aeee;
       }
 
       .hljs-built_in,
       .hljs-title.class_,
       .hljs-class .hljs-title {
-        color: #c18401;
+        color: #e6c07b;
       }
 
       .hljs-emphasis {
