+++ packages/zero-docs/assets/scss/utilities.scss
@@ -59,6 +59,7 @@ $replace4: #000000;
   --divider: #{$alto};
   --link-color: #{$blueRibbon};
   --link-hover-color: #{$melrose};
+  --code__block__color: #{$codGray};
   --code__block__background-color: #{$mineShaft};
   --code__inline__color: #{$codGray};
   --code__inline__background-color: #E4E4E4;
@@ -86,6 +87,7 @@ $replace4: #000000;
   --divider: #{$mineShaft};
   --link-color: #{$dodgerBlue};
   --link-hover-color: #{$bayOfMany};
+  --code__block__color: #{$wildSand};
   --code__block__background-color: #{$mineShaft};
   --code__inline__color: #{$wildSand};
   --code__inline__background-color: #383838;
